+++ FileSpammer/functions/getInput.h
@@ -1,7 +1,7 @@
 #ifndef GETINPUT_H
 #define GETINPUT_H
 
-void getInput(int* spamFiles, int* spamRows, char spamRow[]);
+void getInput(int* spamFiles, long long* spamRows, char spamRow[]);
 int scanfMbAndRows(int maxDigits, char textInLineBeforeNumber[], int spamFiles);
 
 #endif /*GETINPUT_H*/
